+++ ros-20190910/alibabacloud_ros20190910/client.py
@@ -5938,6 +5938,8 @@ class Client(OpenApiClient):
             query['ClientToken'] = request.client_token
         if not UtilClient.is_unset(request.disable_rollback):
             query['DisableRollback'] = request.disable_rollback
+        if not UtilClient.is_unset(request.dry_run):
+            query['DryRun'] = request.dry_run
         if not UtilClient.is_unset(request.parallelism):
             query['Parallelism'] = request.parallelism
         if not UtilClient.is_unset(request.parameters):
@@ -6004,6 +6006,8 @@ class Client(OpenApiClient):
             query['ClientToken'] = request.client_token
         if not UtilClient.is_unset(request.disable_rollback):
             query['DisableRollback'] = request.disable_rollback
+        if not UtilClient.is_unset(request.dry_run):
+            query['DryRun'] = request.dry_run
         if not UtilClient.is_unset(request.parallelism):
             query['Parallelism'] = request.parallelism
         if not UtilClient.is_unset(request.parameters):